course2/sem3/Kotlin/kotlin-2023-5-coroutines-basic/lesson/2-sequential processor/src/SequentialProcessor.kt
import kotlinx.coroutines.DelicateCoroutinesApi
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.newSingleThreadContext
import kotlinx.coroutines.withContext

class SequentialProcessor(private val handler: (String) -> String) : TaskProcessor {
    @OptIn(ExperimentalCoroutinesApi::class, DelicateCoroutinesApi::class)
    private val singleThreadContext = newSingleThreadContext("SequentialProcessorThread")

    override suspend fun process(argument: String): String {
        return withContext(singleThreadContext) {
            handler(argument)
        }
    }
}
